refactor: generalize worker request evaluator
Move the weighted resource request evaluator into `hypha_resources` as a generic `ResourceEvaluator` trait so other crates can leverage the same scoring logic.
