+++ crates/worker/src/lib.rs
@@ -6,8 +6,6 @@ pub mod network;
 
 pub mod arbiter;
 
-pub mod request_evaluator;
-
 pub mod job_manager;
 pub mod lease_manager;
 pub mod resources;
